+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds an /analytics command for server admins.
+
 ## [v18.1.0](https://github.com/lexicalunit/spellbot/releases/tag/v18.1.0) - 2026-02-14
 
 ### Added